IdGeneratorBase → DbIntIdGeneratorBase, SqlServerIdGenerator → SqlServerIntIdGenerator
